backend/ntpip_backend/requirement.txt
Flask==3.1.0
Flask_Cors==4.0.1
flask_jwt_extended==4.7.1
joblib==1.4.2
openmeteo_requests==1.3.0
pandas==2.2.3
pymongo==4.10.1
python-dotenv==1.0.1
Requests==2.32.3
requests_cache==1.2.1
retry_requests==2.0.0
scikit_learn==1.5.2
Werkzeug==3.1.3
openai==1.58.1